Create README.md
Server-side of A-Point-Meant: A-Point-Meant is a scheduling site MVP currently. The current scope is a place to allow a professional to schedule appointments and to-do items. In the future, this site will host a Calendar and allow the admin to post availability that then users can claim and set appointments. 

Project commits:
01-27-22 Initialize project using Create React App, File structure, JWT, React-Calendar started, Models started, Headers, …
01-31-22 Pushed out register endpoint fix
02-01-22 Heroku deploy and tested, DB creation and fixes, appointment model created, appointment endpoint worked on.
02-03-22 Todo table added to DB with message option added.
02-04-22 Fixed appointment date model to make unique, dropped tables and tested
02-06-22 dropped tables to fix Todo relations, added UUID, pushed to Heroku
02-18-22 password and email requirements built-in and tested
02-20-22 Fixed link for full CRUD for appointments and todos
02-21-22 fixed URL links to Heroku URLs